完成在bpmn中动态添加表单-主表单（未完成）

+++ activiti7-java/src/main/java/com/activiti/modules/entity/SysNodeFormEntity.java
@@ -45,7 +45,7 @@ public class SysNodeFormEntity implements Serializable {
     /**
      * 是否是主表单,1:是,2:否
      */
-    private Long isMainFrom;
+    private Integer isMainFrom;
     /**
      * 创建时间
      */
